@@ -5,7 +5,7 @@ import { Student } from '../student';
 @Component({
     selector: 'app-home',
     templateUrl: './home.component.html',
-    styleUrls: ['./home.component.sass']
+    styleUrls: ['./home.component.css']
 })
 export class HomeComponent implements OnInit {
 
