@@ -1,12 +1,15 @@
-import { TestBed } from '@angular/core/testing';
+import { TestBed, inject } from '@angular/core/testing';
 
 import { StudentService } from './student.service';
 
 describe('StudentService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: StudentService = TestBed.get(StudentService);
-    expect(service).toBeTruthy();
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [StudentService]
+    });
   });
+
+  it('should be created', inject([StudentService], (service: StudentService) => {
+    expect(service).toBeTruthy();
+  }));
 });
