@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Student } from './student';
 import { Observable } from 'rxjs';
-import { ActionStatus } from './action-status';
-
+import { Student } from './student';
 
 const API_URL: string = 'http://localhost:8081/students';
 const GET_ALL_STUDENTS_URL: string = API_URL + '/all';
@@ -33,5 +31,5 @@ export class StudentService {
     getAllStudents(): Observable<Student[]> {
         return this.http.get<Student[]>(GET_ALL_STUDENTS_URL);
     }
-
+    
 }
